Make callback more flexible and easy to implement.
Remove boilerplate about when to run callback and provide as a facility
in the wrapper.

Allow behavior of update() to be statically polymorphic with templating
on the existence of PotentialT::callback.

To do: this is laid out wrong and should just call the member function
directly.

To do: template on multiple callback signatures depending on whether
ensemble resources are needed.