gmxapi-62 Stop condition hook (work in progress)
This is mostly what the updates for modular stop condition code will
look like in the short term. There will be another update to specify
the name of the output to be bound. The code should be effective as
a model and should build against the corresponding commit in
gromacs-gmxapi, but the newly described facility will not yet have an
effect.