gmxapi-62 Stop condition hook (work in progress)
Add some checks for initialized objects. Note that an uninitialized
py::object apparently has a sensible implicit bool conversion that we
can use.

To do: we need to make sure to provide some warning or stronger error
if we don't find the subscribers we are expecting. Right now, we will
just silently attach a no-op to the workflow in some circumstances.