Allow tests to tolerate unclean working directories.
Newer code bases have a problem with multiple tests run in the same
working directory. The symptom is that the GROMACS simulation will shut
down cleanly, but with an exit code of 1, and console output that
includes:

"Fatal error: Checksum wrong for 'md.log'. The file has been replaced or
its contents have been modified. Cannot do appending because of this
condition."

The pytest "cleandir" wrapper should have taken care of this, and will
be addressed in a separate change. But reusing the working directories
in these particular tests is not a problem, so this commit instructs
the test simulations not to try to append output to existing files.